Add Labels.asMap and hide the builder implementation.
